+++ src/Autocomplete/Autocomplete.stories.tsx
@@ -20,6 +20,7 @@ import CaretDownIcon from '@hcl-software/enchanted-icons/dist/carbon/es/caret--d
 import SearchIcon from '@hcl-software/enchanted-icons/dist/carbon/es/search';
 import InformationIcon from '@hcl-software/enchanted-icons/dist/carbon/es/information';
 import WarningIcon from '@hcl-software/enchanted-icons/dist/carbon/es/warning--alt';
+import PopupIcon from '@hcl-software/enchanted-icons/dist/carbon/es/popup';
 
 import { Alert, Box, SvgIconProps } from '@mui/material';
 import Autocomplete from './Autocomplete';
@@ -29,6 +30,7 @@ import ListItemText from '../List/ListItemText';
 import { TooltipPlacement } from '../Tooltip';
 import CircularProgress from '../ProgressIndicator/CircularProgress';
 import Typography from '../Typography';
+import IconButton from '../IconButton';
 
 export default {
   title: 'Inputs/Autocomplete',
@@ -191,6 +193,10 @@ export default {
     endAdornmentAction: {
       control: false,
       description: 'Action of the end endAdornment icon.',
+    },
+    endAdornmentIconButton: {
+      control: false,
+      description: 'IconButton of the end endAdornment icon.',
     },
     renderNonEditInput: {
       control: false,
@@ -497,5 +503,22 @@ export const ExampleAutocompleteWithCustomizedBanner = {
         story: 'Autocomplete with warning only - no errors. Shows how to display warnings about content availability without blocking user interaction.',
       },
     },
+  },
+};
+
+export const ExampleAutocompleteWithEndAdornmentIconButton = {
+  render: Template,
+  args: {
+    ...ExampleAutocomplete.args,
+    value: top100Films[0],
+    error: true,
+    freeSolo: false,
+    startAdornment: 'SearchIcon',
+    endAdornment: 'Loading',
+    endAdornmentIconButton: (
+      <IconButton>
+        <PopupIcon style={{ width: 16, height: 16 }} />
+      </IconButton>
+    ),
   },
 };